refactor(cloud): share PostgreSQL schema startup between services
